format spring-tx-scenario
