+++ test/plugin/scenarios/spring-tx-scenario/src/main/java/test/apache/skywalking/apm/testcase/spring/transaction/Application.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.spring.transaction;
+package test.apache.skywalking.apm.testcase.spring.transaction;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
