+++ test/plugin/scenarios/spring-tx-scenario/src/main/java/test/apache/skywalking/apm/testcase/spring/transaction/controller/CaseController.java
@@ -16,7 +16,7 @@
  *
  */
 
-package test.org.apache.skywalking.apm.testcase.spring.transaction.controller;
+package test.apache.skywalking.apm.testcase.spring.transaction.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -25,7 +25,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
-import test.org.apache.skywalking.apm.testcase.spring.transaction.service.DemoService;
+import test.apache.skywalking.apm.testcase.spring.transaction.service.DemoService;
 
 @Controller
 @RequestMapping("/case")
